feat(testing): suíte de testes de arquitetura e prevenção de regressão no app core (#300)
- Adiciona BaseTenantIsolationTest e testes parametrizados para os modelos de weddings, finances, logistics e scheduler.
- Adiciona teste de arquitetura de API Ninja garantindo operation_id em 100% dos endpoints e 401 para rotas desprotegidas.
- Adiciona testes de rollback de transação atômica simulando falhas em escritas multi-modelo.
- Adiciona testes de performance de listagem validando complexidade de queries O(1) e prevenção de N+1.
- Adiciona auditoria estática por AST proibindo o uso de django.shortcuts.get_object_or_404 e exigindo o parâmetro company em serviços.
- Adiciona inspeção de Schemas Pydantic garantindo o uso de Decimal para campos monetários (BR-VAL01).
- Adiciona validação de travas de concorrência (.select_for_update()) em operações críticas.
- Expande testes do BaseModel para 100% de conformidade com a ADR-011.

+++ backend/apps/weddings/schemas.py
@@ -1,4 +1,5 @@
 import datetime
+from decimal import Decimal
 from enum import StrEnum
 
 from ninja import Field, Schema
@@ -44,7 +45,7 @@ class WeddingOut(Schema):
     template: str | None
     created_at: datetime.datetime
     updated_at: datetime.datetime
-    total_budget: float | None = Field(None, ge=0)
+    total_budget: Decimal | None = Field(None, ge=0)
     overdue_installments: int = Field(0, ge=0)
     incomplete_tasks: int = Field(0, ge=0)
